Split tracked PR status comment boundaries
