+++ src/family-directory-layout.test.ts
@@ -72,6 +72,9 @@ const EXPECTED_TOP_LEVEL_ENTRIES = {
     "tracked-pr-lifecycle-projection.ts",
     "tracked-pr-local-ci-publication-gate.ts",
     "tracked-pr-ready-promotion-blocker.ts",
+    "tracked-pr-status-comment-marker.ts",
+    "tracked-pr-status-comment-publisher.ts",
+    "tracked-pr-status-comment-rendering.ts",
     "tracked-pr-status-comment.ts",
     "turn-execution-failure-helpers.ts",
     "turn-execution-orchestration.ts",
